+++ CloudHarvestCoreTasks/__register__.py
@@ -5,18 +5,4 @@ kicks off the @register_definition decorator, which registers all the classes wi
 to ensure that the TaskRegistry is populated with all the classes that are available to the system.
 """
 
-from .chains import (
-    ReportTaskChain
-)
-
-from .tasks import (
-    DummyTask,
-    ErrorTask,
-    FileTask,
-    ForEachTask,
-    HarvestRecordSetTask,
-    MongoTask,
-    PruneTask,
-    RedisTask,
-    WaitTask
-)
+from tasks import *